Fixed model bugs and moved Review to courses app (#99)
* Fixed model bugs and moved Review to courses app

* Added gitignore update for migration folders
