+++ ratenyu/courses/models.py
@@ -1,9 +1,12 @@
+import datetime
+
 from django.db import models
+from django.utils import timezone
 
-from ratenyu.professors.models import Professor
+from professors.models import Professor
 
 
-class Courses(models.Model):
+class Course(models.Model):
     course_id = models.CharField(max_length=250, primary_key=True)
     course_title = models.CharField(max_length=250)
     course_subject_code = models.CharField(max_length=20, blank=True)
@@ -13,10 +16,11 @@ class Courses(models.Model):
     def __str__(self):
         return self
 
+
 class Class(models.Model):
     class_id = models.CharField(max_length=250, primary_key=True)
     professor = models.ForeignKey(to=Professor, on_delete=models.CASCADE)
-    course = models.ForeignKey(to=Courses, on_delete=models.CASCADE)
+    course = models.ForeignKey(to=Course, on_delete=models.CASCADE)
     class_type = models.CharField(max_length=250, blank=True)
     class_section = models.CharField(max_length=250, blank=True)
     term = models.CharField(max_length=250, blank=True)
@@ -26,3 +30,23 @@ class Class(models.Model):
 
     def __str__(self):
         return self
+
+
+class Review(models.Model):
+    review_text = models.TextField()
+    rating = models.PositiveIntegerField(choices=(
+        (1, '1'),
+        (2, '2'),
+        (3, '3'),
+        (4, '4'),
+        (5, '5')
+    ))
+    class_id = models.ForeignKey(to=Class, on_delete=models.CASCADE)
+    user = models.CharField(max_length=250)
+    pub_date = models.DateTimeField('date published')
+
+    def __str__(self):
+        return self.review_text
+
+    def was_published_recently(self):
+        return self.pub_date >= timezone.now() - datetime.timedelta(days=1)
