+++ ratenyu/professors/models.py
@@ -1,9 +1,4 @@
-import datetime
-
 from django.db import models
-from django.utils import timezone
-
-from ratenyu.courses.models import Class
 
 
 class Professor(models.Model):
@@ -14,16 +9,3 @@ class Professor(models.Model):
 
     def __str__(self):
         return self
-
-class Review(models.Model):
-    review_text = models.TextField()
-    rating = models.PositiveIntegerField(choices=[1, 2, 3, 4, 5])
-    class_id = models.ForeignKey(to=Class, on_delete=models.CASCADE)
-    user = models.CharField(max_length=250)
-    pub_date = models.DateTimeField('date published')
-    
-    def __str__(self):
-        return self.reviewText
-    
-    def was_published_recently(self):
-        return self.pub_date >= timezone.now() - datetime.timedelta(days=1)
